mint2g: init: Define pty_symlink as a fake service
phoneserver_2sim, being sprd's internal aspects and strictly closed source, cannot be modified. And it starts the pty_symlink service which creates the AT channel sockets.
Now for some reason in nougat, pty_symlink is unable to run.
If we undefine the service completely, dmesg will be spammed by "No service called pty_symlink" like messages as phoneserver will constantly try to run the service. If it is properly defined, the service will be killed by init, and debugging the reason for the same is same as understanding a girl's mind.
Hence define it as a fake service, which on execution will do nothing. As a result, phoneserver will think it was able to run the service, and keep the dmesg green.

Signed-off-by: corphish <d97.avinaba@gmail.com>